MyApps-13: Move paging implementation, api and dependency injection to submodules.
