--- app/src/main/java/xyz/torquato/myapps/data/impl/paging/model/PageRequest.kt
@@ -1,8 +1,0 @@
-package xyz.torquato.myapps.data.impl.paging.model
-
-sealed interface PageRequest {
-
-    data class Fixed(
-        val size: Int
-    ): PageRequest
-}
