+++ app/src/main/java/xyz/torquato/myapps/di/data/DataModule.kt
@@ -5,7 +5,6 @@ import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import xyz.torquato.myapps.api.math.IMathRepository
-import xyz.torquato.myapps.api.pagging.IPagingRepository
 import xyz.torquato.myapps.api.sound.ICypherRepository
 import xyz.torquato.myapps.api.sound.ISoundRepository
 import xyz.torquato.myapps.api.web.ILocalWebRepository
@@ -13,7 +12,6 @@ import xyz.torquato.myapps.api.web.IQueryRepository
 import xyz.torquato.myapps.api.web.IWebRepository
 import xyz.torquato.myapps.api.web.cache.IWebCacheRepository
 import xyz.torquato.myapps.data.cypher.CypherRepository
-import xyz.torquato.myapps.data.impl.paging.PagingRepository
 import xyz.torquato.myapps.data.math.MathRepository
 import xyz.torquato.myapps.data.waves.SoundRepository
 import xyz.torquato.myapps.data.web.LocalWebRepository
@@ -52,9 +50,5 @@ abstract class RepositoryModule {
     @Binds
     @Singleton
     abstract fun bindLocalWebImpl(impl: LocalWebRepository): ILocalWebRepository
-
-    @Binds
-    @Singleton
-    abstract fun bindPagingImpl(impl: PagingRepository): IPagingRepository
 
 }
