+++ app/src/main/java/xyz/torquato/myapps/di/domain/DomainWebModule.kt
@@ -4,7 +4,6 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import xyz.torquato.myapps.api.pagging.IPagingRepository
 import xyz.torquato.myapps.api.web.ILocalWebRepository
 import xyz.torquato.myapps.api.web.IQueryRepository
 import xyz.torquato.myapps.api.web.IWebRepository
@@ -14,6 +13,7 @@ import xyz.torquato.myapps.domain.web.SetSelectedBookUseCase
 import xyz.torquato.myapps.domain.web.content.GetBooksUseCase
 import xyz.torquato.myapps.domain.web.content.GetCacheBooksUseCase
 import xyz.torquato.myapps.domain.web.pagging.GetNextPageUseCase
+import xyz.torquato.myaps.domain.api.pagging.IPagingRepository
 import javax.inject.Singleton
 
 @Module
