+++ app/src/main/java/xyz/torquato/myapps/domain/web/pagging/RemovePagerUseCase.kt
@@ -1,6 +1,6 @@
 package xyz.torquato.myapps.domain.web.pagging
 
-import xyz.torquato.myapps.api.pagging.IPagingRepository
+import xyz.torquato.myaps.domain.api.pagging.IPagingRepository
 
 class RemovePagerUseCase(
     private val repository: IPagingRepository
