+++ gradle/libs.versions.toml
@@ -27,6 +27,8 @@ openssl = "1.1.1q-beta-1"
 jsoncpp = "1.9.5-beta-1"
 gtest = "1.11.0-beta-1"
 jetbrainsKotlinJvm = "2.0.20"
+coroutines-core ="1.10.2"
+
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
@@ -64,6 +66,8 @@ thirdparty-gtest = { group = "com.android.ndk.thirdparty", name = "googletest", 
 
 glide = { group = "com.github.bumptech.glide", name = "glide", version.ref = "glide"}
 glide-compose = { group = "com.github.bumptech.glide", name = "compose", version.ref = "glide-compose"}
+
+corountines-core = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-core", version.ref = "coroutines-core"}
 
 [plugins]
 android-application = { id = "com.android.application", version.ref = "agp" }
