+++ settings.gradle.kts
@@ -26,3 +26,4 @@ include(":domain:api")
 include(":data:impl")
 include(":domain:di")
 include(":data:di")
+include(":data:di")
